cpp08 fix y cpp09

+++ cpp08/ex01/Span.hpp
@@ -30,8 +30,8 @@ public:
 		_data.insert(_data.end(), begin, end);
 	}
 
-	int	shortestSpan() const;
-	int	longestSpan() const;
+	long	shortestSpan() const;
+	long	longestSpan() const;
 };
 
 #endif
